pyproject.toml
[project]
name = "adk-pipe"
version = "0.1.0"
description = ""
authors = [
    {name = "Jordan",email = "jordantotten@google.com"}
]
readme = "README.md"

requires-python = ">=3.13, <4.0"
dependencies = [
    "google-genai>=1.19.0,<2.0.0",
    "google-cloud-aiplatform[agent-engines]>=1.125.0,<2.0.0",
    "google-adk[eval]>=1.31.1",
    "pandas>=2.3.2,<3.0.0",
    "google-cloud-bigquery>=3.36.0,<4.0.0",
    "db-dtypes>=1.4.3,<2.0.0",
    "tabulate>=0.9.0,<0.10.0",
    "uvicorn>=0.35.0",
    "fastapi>=0.124.1",
    "pillow>=11.3.0,<12.0.0",
    "markdown>=3.8.2,<4.0.0",
    "markdown-pdf>=1.9,<2.0",
    "absl-py>=2.3.1,<3.0.0",
    "functions-framework==3.9.2",
    "cloudevents==1.11.0",
    "google-cloud-pubsub>=2.31.1,<3.0.0",
    "litellm>=1.83.7",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
